update Orders formatting and one object per class
