+++ code/Order/Payment_ProcessorExtension.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SwipeStripe\Core\Order;
+
+use SilverStripe\Core\Extension;
+
+class Payment_ProcessorExtension extends Extension
+{
+    public function onBeforeRedirect()
+    {
+        $order = $this->owner->payment->Order();
+        if ($order && $order->exists()) {
+            $order->onAfterPayment();
+        }
+    }
+}
